Fix YAMLLINT
As part of this PR, YAMLLint was enabled. RUAMEL does not dump YAMLLint
valid YAML in some cases (e.g. a list of lists). A file with specific
YAML syntax was created to represent the data which RUAMEL should dump,
and a unit test was written which asserts that the data RUAMEL dumps as a
string exactly matches the string in this file.

When this file is evaluated against YAMLLint standards, it fails. If modified
for compliance, when the unit test tries to assert that the RUAMEL data it dumps
is valid, it fails. I tried to modify RUAMEL's dumping logic in order to write it
in a YAMLLint compliant format, but was unable to do so. Because of this, I was
forced to disable YAMLLint's 'hyphens' rule for the code block in question. I then
had to have the unit test strip the comments denoting the disable of the YAMLLint
rule.